+++ src/main/java/oi/projet/springboot/ImmobilierApp/Services/LocataireService.java
@@ -1,13 +1,10 @@
 package oi.projet.springboot.ImmobilierApp.Services;
 
-import java.util.ArrayList;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import jakarta.persistence.EntityNotFoundException;
 import oi.projet.springboot.ImmobilierApp.models.Locataire;
 import oi.projet.springboot.ImmobilierApp.repository.LocataireRepository;
 
@@ -17,44 +14,23 @@ public class LocataireService {
     @Autowired
     private LocataireRepository locataireRepository;
 
-    public List<Locataire> getLocataires() {
-        List<Locataire> locataires = new ArrayList<>();
-        locataireRepository.findAll().forEach(locataires::add);
-        return locataires;
+    public List<Locataire> getAllLocataires() {
+        return locataireRepository.findAll();
     }
 
-    public Locataire getUnLocataire(long id) {
-        return locataireRepository.findById(id).orElse(null);
+    public Optional<Locataire> getLocataireById(Long id) {
+        return locataireRepository.findById(id);
     }
 
-    public void deleteLocataire(long id) {
-        locataireRepository.deleteById(id);
+    public Locataire saveLocataire(Locataire locataire) {
+        return locataireRepository.save(locataire);
     }
 
-    public void addLocataire(Locataire locataire) {
-        locataireRepository.save(locataire);
-    }
-
-    public void updateLocataire(Locataire locataire, long id) {
-        Optional<Locataire> existingLocataireOptional = locataireRepository.findById(id);
-
-        if (existingLocataireOptional.isPresent()) {
-            Locataire existingLocataire = existingLocataireOptional.get();
-            // Mettez à jour les champs nécessaires de l'entité existante
-            existingLocataire.setImage(locataire.getImage());
-            existingLocataire.setNom(locataire.getNom());
-            existingLocataire.setPrenom(locataire.getPrenom());
-            existingLocataire.setAdresse(locataire.getAdresse());
-            existingLocataire.setEmail(locataire.getEmail());
-            existingLocataire.setContact(locataire.getContact());
-            existingLocataire.setNomUtilisateur(locataire.getNomUtilisateur());
-            existingLocataire.setMotDePasse(locataire.getMotDePasse());
-
-            // Sauvegardez l'entité mise à jour
-            locataireRepository.save(existingLocataire);
+    public void deleteLocataire(Long id) {
+        if (locataireRepository.existsById(id)) {
+            locataireRepository.deleteById(id);
         } else {
-            // Gérer le cas où l'entité n'existe pas (par exemple, lever une exception)
-            throw new EntityNotFoundException("Locataire with id " + id + " not found.");
+            throw new RuntimeException("Locataire introuvable avec l'ID : " + id);
         }
     }
 }
